ultioms ajustes front nagular panel admin

+++ front-angular/src/app/usuarios/index/index.component.ts
@@ -4,6 +4,7 @@ import { Router } from '@angular/router';
 import { UsuarioService } from '../../servicios/usuario.service';
 import { Cafeteria } from '../../modelos/cafeteria.model';
 import { Admin } from '../../modelos/admin.model';
+import { finalize } from 'rxjs';
 
 @Component({
   selector: 'app-index',
@@ -18,6 +19,9 @@ export class IndexComponent {
   clave: string | null = null;
   listaCafeterias : Cafeteria[] = [];
   listaAdmins : Admin[] = [];
+
+  noEliminarCafeteria: boolean = false;
+  noEliminarAdmin: boolean = false;
 
   constructor(private usuarioService : UsuarioService, private _router : Router){
   }
@@ -63,13 +67,37 @@ export class IndexComponent {
 
   eliminarCafeteria(id:any):void {
     this.usuarioService.deleteCafeteria(this.clave, id).subscribe(
-      data => { }, err => { console.log(err); }
+      data => { 
+        this.noEliminarCafeteria = false;
+        this.cargarCafeterias();
+      }, err => { 
+        if (err.status) {
+          this.noEliminarCafeteria = true;
+        }
+        console.log(err); 
+      }
     )
   }
 
   eliminarAdmin(id:any):void {
-    this.usuarioService.deleteAdmin(this.clave, id).subscribe(
-      data => { this.cargarAdmins() }, err => { console.log(err); }
-    )
+    var user_id = localStorage.getItem('user_id');
+    if (user_id != id) {
+      this.usuarioService.deleteAdmin(this.clave, id).subscribe(
+        data => {
+          this.noEliminarAdmin = false;
+          this.cargarAdmins();
+        }, err => { 
+
+          console.log(err); 
+        }
+      )
+    }
+    else {
+      this.noEliminarAdmin = true;      
+
+      console.log("No se puede elmimar");
+      
+    }
+    
   }
 }
